move UserInfoOfDay component to another file
